add swagger and bcrypt

+++ controllers/OtpController.js
@@ -1,5 +1,6 @@
 const moment = require('moment')
 const mongoose = require('mongoose')
+const bcrypt = require('bcrypt')
 const bot = require('../bot')
 
 const { UserChat, Otp } = require('../models')
@@ -20,33 +21,37 @@ exports.generate = async (req, res) => {
   }
 
   const otpVal = Math.floor(Math.random() * (999999 - 100000) + 100000)
+
+  const salt = bcrypt.genSaltSync(12)
+  const hashOtp = bcrypt.hashSync(otpVal.toString(), salt)
+
   const exp = moment(new Date()).add(5, 'minutes')
 
   const otp = await Otp.findOne({ user_id: userId })
   if (otp) {
     if (moment(otp.exp).subtract(270, 'seconds').isAfter(new Date())) {
-      res.send({ success: false })
+      res.status(400).json({ success: false })
       return
     }
-    otp.otp = otpVal
+    otp.otp = hashOtp
     otp.exp = exp
     await otp.save()
   } else {
     const otpNew = new Otp({
       _id: new mongoose.Types.ObjectId(),
       user_id: userId,
-      otp: otpVal,
+      otp: hashOtp,
       exp
     })
     await otpNew.save()
   }
 
-  bot.telegram.sendMessage(userChat.chat_id, `Your OTP Code: ${otpVal}`)
+  bot.telegram.sendMessage(userChat.chat_id, `Your One Time Password is ${otpVal}`)
     .then(() => {
       res.send({ success: true })
     })
     .catch(() => {
-      res.send({ success: false })
+      res.status(400).json({ success: false })
     })
 }
 
@@ -62,15 +67,19 @@ exports.verify = async (req, res) => {
     return
   }
 
-  const otp = await Otp.findOne({ user_id: userId, otp: otpVal })
+  const otp = await Otp.findOne({ user_id: userId })
   if (otp) {
-    await otp.delete()
-    if (moment(otp.exp).isBefore(new Date())) {
-      res.send({ success: false })
+    if (bcrypt.compareSync(req.body.otp, otp.otp)) {
+      await otp.delete()
+      if (moment(otp.exp).isBefore(new Date())) {
+        res.status(400).json({ success: false })
+      } else {
+        res.send({ success: true })
+      }
     } else {
-      res.send({ success: true })
+      res.status(400).json({ success: false })
     }
   } else {
-    res.send({ success: false })
+    res.status(400).json({ success: false })
   }
 }
